Added multi motor and multi directional support
This update essentially makes the bot functional. In this version, I have added another class for a second motor, a variable to hand direction, and a switch statement within one of the if's. I also removed the first moveTo statement within the void setup which actually increased the speed. Within the if responsible for handling the spam from the IR remote, I added the switch statement to handle direction. I also found that the steppers would not respond to directional code unless they were first accelerated at half step. Basically, depending on which button is pressed, a if statement primes the directional variable with a value. Since the remote is programmed to send the same spam response after any button has been pressed, I use the value in the directional variable to tell me what has been pressed. Let it be noted that one motor must go in the opposite direction when going backward and forward in order for both wheels to move in the same direction. Any further commits to this code would be to improve effeciency. Otherwise it is fully functional with perhaps a need for a few more comments...